+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### Devop
 
+* Fix hardware access - network and address select  [#2937](https://github.com/MyEtherWallet/MyEtherWallet/pull/2937)
 * Add MEWconnect [#2934](https://github.com/MyEtherWallet/MyEtherWallet/pull/2934)
 * Fix send and sign tx [#2929](https://github.com/MyEtherWallet/MyEtherWallet/pull/2929)
 * Merge UI fixes [#2916](https://github.com/MyEtherWallet/MyEtherWallet/pull/2916)